backend/requirements-auth.txt
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.0
mysql-connector-python==8.0.26
sqlalchemy==1.4.23 